backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
langgraph==0.6.7
pydantic==2.11.7
python-multipart==0.0.20
python-docx==1.2.0
python-pptx==1.0.2
